refactor: reorganize and update imports for `DeleteUserService` and `EditUserCommand` components
- Moved `DTO`, `Command`, and `Validator` files to appropriate `Application` and `Domain` subdirectories.
- Adjusted import paths in corresponding services and handlers.
