+++ user-service/src/Domain/Command/Handlers/EditUserCommandHandler.ts
@@ -2,9 +2,9 @@ import {BaseHandlerCommand} from "./BaseHandlerCommand.js";
 import {EditUserCommand} from "../CommandObject/EditUserCommand.js";
 import {UserRepository} from "../../../Infrastructure/Persistence/Repositories/Concrete/UserRepository.js";
 import {NotificationError} from "../../../Shared/Errors/NotificationError.js";
-import {PasswordHashVO} from "../../../Domain/ValueObjects/PasswordHashVO.js";
-import {EmailVO} from "../../../Domain/ValueObjects/EmailVO.js";
-import {User} from "../../../Domain/Entities/Concrete/User.js";
+import {PasswordHashVO} from "../../ValueObjects/PasswordHashVO.js";
+import {EmailVO} from "../../ValueObjects/EmailVO.js";
+import {User} from "../../Entities/Concrete/User.js";
 
 export class EditUserCommandHandler implements BaseHandlerCommand<EditUserCommand>
 {
