+++ user-service/src/Domain/Command/Validators/EditUserCommandValidator.ts
@@ -2,9 +2,9 @@ import {BaseValidator} from "./BaseValidator.js";
 import {EditUserCommand} from "../CommandObject/EditUserCommand.js";
 import {UserRepository} from "../../../Infrastructure/Persistence/Repositories/Concrete/UserRepository.js";
 import {NotificationError} from "../../../Shared/Errors/NotificationError.js";
-import {EmailVO} from "../../../Domain/ValueObjects/EmailVO.js";
+import {EmailVO} from "../../ValueObjects/EmailVO.js";
 import {ErrorCatalog} from "../../../Shared/Errors/ErrorCatalog.js";
-import {PasswordHashVO} from "../../../Domain/ValueObjects/PasswordHashVO.js";
+import {PasswordHashVO} from "../../ValueObjects/PasswordHashVO.js";
 import {CustomError} from "../../../Shared/Errors/CustomError.js";
 import {ValidationException} from "../../../Shared/Errors/ValidationException.js";
